fix bug move server.js to utils
