+++ client/app/utils/credentials.js
